Fix Fan Multiplexer init
